refactor: extract the delta-row builder to its own module
The content-edit path has to emit its delta inside the reversal transaction, which applyDeltaAction cannot do: it owns its own transaction, and its barrier rejects a user_edit action while a reversal is in flight. Pulling the row shape and the MAX+1 log-position rule out keeps one home for both rather than growing a second copy in a domain file where nobody would look for it. No behavior change — apply-delta-action's tests pass unmodified.
